Mini-Vision-Transformer-on-Tiny-Imagenet/requirements.txt
flask
flask-cors 
torch 
torchvision 
pandas 
pillow
requests
matplotlib